Add new navbar to pdfviewer, move addEnsemble() to pers and replace the html string with a template
