+++ templates/web/desktop.html
@@ -9,6 +9,7 @@
   -->
     <title>NB Desktop</title>
     <link type="text/css" href="/content/compiled/desktop.css" rel="stylesheet" />
+    <link type="text/css" href="/content/lib/font-awesome/font-awesome-4.6.1/css/font-awesome.min.css" rel="stylesheet" />
     <script type="text/javascript" src="/content/compiled/desktop_NB.js"></script>
   </head>
   <body/>
