+++ templates/web/source1.html
@@ -9,6 +9,7 @@
   -->
     <title>NB File Viewer</title>
     <link type="text/css" href="/content/compiled/pdfviewer.css" rel="stylesheet" />
+    <link type="text/css" href="/content/lib/font-awesome/font-awesome-4.6.1/css/font-awesome.min.css" rel="stylesheet" />
     <script type="text/javascript" src="/content/compiled/pdfviewer_NB.js"></script>
     <link href="http://code.jquery.com/ui/1.9.2/themes/smoothness/jquery-ui.css" type="text.css"/ rel="stylesheet">
   </head>
